Implemented Rating Functionality

+++ MVC/CIPlatform/Entities/ViewModels/MissionList.cs
@@ -23,9 +23,11 @@ namespace Entities.ViewModels
         public string? missionImage { get; set; }
         public string? mediaPath { get; set; }
         public string? mediaType { get; set; }
-        public int? Rating { get; set; }
+        public decimal? Rating { get; set; }
         public int? goalValue { get; set; }
         public DateTime? deadline { get; set; }
         public long? favMissionId { get; set; }
+
+
     }
 }
